move parseLengthAndMagic into separate file
